backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
redis==5.0.1
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
